Add polymorphic input check for add, move all files to duke package
