+++ src/main/java/duke/DukeException.java
@@ -1,5 +1,7 @@
+package duke;
+
 public class DukeException extends RuntimeException {
     public DukeException(String message) {
-        super(message);
+        super(Duke.styleResponse("Oops, " + message));
     }
 }
