+++ src/main/java/duke/Item.java
@@ -1,6 +1,12 @@
+package duke;
+
 abstract class Item {
     private String name;
     private boolean isDone = false;
+
+    public Item(String[] strings) {
+        this.checkInput(strings);
+    }
 
     public void markAsDone() {
         this.isDone = true;
@@ -8,6 +14,12 @@ abstract class Item {
 
     public void setName(String name) {
         this.name = name;
+    }
+
+    public void checkInput(String[] strings) {
+        if (strings.length == 1) {
+            throw new DukeException("I need more than just the command");
+        }
     }
 
     @Override
