+++ src/main/java/duke/ToDo.java
@@ -1,6 +1,9 @@
+package duke;
+
 public class ToDo extends Item {
 
     public ToDo(String[] strings) {
+        super(strings);
         String line = String.join(" ", strings);
         this.setName(line.substring(5, line.length()));
     }
